chore(tests): share the mock TimerImpl across qsys test TUs
EventManager owns a single process-wide TimerImpl and asserts on a second
initTimer(), so a test TU cannot install its own mock. Extract the static
shared-instance pattern from test_animmgr.cpp into MockTimerImpl.hpp so
other TUs in the test_qsys binary can drive the same controllable clock.

No behavior change.
